pyproject.toml
[project]
name = "flaskv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "mysql-connector-python>=9.4.0",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "waitress>=3.0.2",
]
